% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_wcid}
\alias{align_wcid}
\title{Set wc_identifier name to releases or diversions}
\usage{
align_wcid(x = NULL)
}
\arguments{
\item{x}{character indicating whether "diversion" or "release" should be returned. Defaults to NULL and thus "diversion"}
}
\value{
wc_identifier equaling either "diversion" or "release"
}
\description{
Internal function for getting correct wc_identifier code for querying
}
