% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aoi_error_msg}
\alias{aoi_error_msg}
\title{Error message handling for extract_coords function}
\usage{
aoi_error_msg()
}
\value{
character error message
}
\description{
Internal helper function that returns a boilerplate error message used in extract_coords function
}
