% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aoi_mask}
\alias{aoi_mask}
\title{Mask points returned from CDSS to only those within polygon AOI}
\usage{
aoi_mask(aoi = NULL, pts = NULL)
}
\arguments{
\item{aoi}{list of length 2 containing an XY coordinate pair, 2 column matrix/dataframe of XY coordinates, sf or Terra SpatVector point/polygon/linestring geometry. Default is NULL.}

\item{pts}{dataframe of points that should be masked to the given aoi. Dataframe must contain "latitude" and "longitude"columns}
}
\value{
dataframe containing subset (or original) point data from the CDSS API
}
\description{
Internal function for masking out extraneous points that fall outside of bounds of provided 'aoi' polygon. If not 'aoi' is provided to the function, the default behavior is to return the original set of points received from CDSS. Furthermore, if the AOI is anything other than a sf or terra polygon, the function will return the original set of data from CDSS.
}
