% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoints_data.R
\docType{data}
\name{api_endpoints}
\alias{api_endpoints}
\title{List of CDSS API endpoints}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61 rows and 5 columns.
}
\usage{
api_endpoints
}
\description{
Dataframe containing a list of all CDSS REST API endpoints
}
\examples{
\dontrun{
 cdssr::api_endpoints
}

}
\keyword{datasets}
