% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_api.R
\name{browse_api}
\alias{browse_api}
\title{Locate all CDSS API Endpoints}
\usage{
browse_api()
}
\value{
dataframe with API endpoint names, URLs and descriptions of each CDSS resource
}
\description{
Returns a dataframe with the API endpoints for CDSS REST services
}
