% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args}
\alias{check_args}
\title{Check all arguments of a function for any/all NULL values}
\usage{
check_args(arg_lst = NULL, ignore = NULL, f = "any")
}
\arguments{
\item{arg_lst}{list of function arguments by calling 'as.list(environment())'}

\item{ignore}{character vector of function arguments to ignore NULL check}

\item{f}{character, either "any" or "all" to indicate whether to check for "any" or "all" NULL argument. If "any" then if any of the function arguments are NULL, then an error is thrown. If "all" then all relevant arguments must be NULL for an error to be thrown. Defaults to "any"}
}
\value{
character error statement with NULL arguments listed, or NULL if no error is thrown by NULL values
}
\description{
Internal utils function for checking a function arguments for any/all invalid/missing arguments necessary to the function it is called within
}
