% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_radius}
\alias{check_radius}
\title{Validate radius for location search queries}
\usage{
check_radius(aoi, radius)
}
\arguments{
\item{aoi}{list of length 2 containing an XY coordinate pair, 2 column matrix/dataframe of XY coordinates, sf or Terra SpatVector point/polygon/linestring geometry}

\item{radius}{numeric, search radius in miles around given point (or the centroid of a polygon). If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}
}
\value{
numeric radius value or NULL
}
\description{
Internal helper function. Checks that an AOI is given and that the radius is within valid range of values. Radius must be: 0 < radius < 150.
}
