% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_vect}
\alias{collapse_vect}
\title{Collapse a list/vector with a given separator into a single string}
\usage{
collapse_vect(x = NULL, sep = "\%2C+")
}
\arguments{
\item{x}{list, or vector to collapse into single string}

\item{sep}{character to separate list/vector items by}
}
\value{
single character string separated by given sep arg
}
\description{
Internal function for generating query string URls
}
