% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract lat/long coordinates from an AOI}
\usage{
extract_coords(aoi, verbose = FALSE)
}
\arguments{
\item{aoi}{2 column matrix/dataframe of XY coordinates, or SF point or polygon object to search for administrative structures within a given radius}

\item{verbose}{logical, whether to print messages to console. Default is FALSE, no print.}
}
\value{
dataframe with lat/long columns
}
\description{
Internal helper function. Checks that an AOI is of valid type and then extracts the latitude/longitude of the point or centroid of the given polygon.
}
