% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_calls.R
\name{get_admin_calls}
\alias{get_admin_calls}
\title{Request active/historic administrative calls data}
\usage{
get_admin_calls(
  active = TRUE,
  min_division = NULL,
  max_division = NULL,
  location_wdid = NULL,
  call_number = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{active}{logical, whether to get active or historical administrative calls. Default iS TRUE and will retrieve active administrative calls.}

\item{min_division}{numeric, minimum division number to get administrative calls for. Default is 1.}

\item{max_division}{numeric, maximum division number to get administrative calls for. Default is 7.}

\item{location_wdid}{numeric, call location structure WDID}

\item{call_number}{numeric, unique call identifier}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of administrative calls
}
\description{
Request active/historic administrative calls data
}
\examples{
# Retrieve ACTIVE administrative calls
active_calls <- get_admin_calls(
               active        = TRUE,
               min_division  = 1,
           )
active_calls

# Retrieve HISTORICAL administrative calls
historic_calls <- get_admin_calls(
               active         = FALSE,
               min_division = 1,
               max_division = 7,
               start_date   = "2000-01-01",
               end_date     = "2005-01-01"
               )
historic_calls
}
