% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysisservices.R
\name{get_call_analysis_wdid}
\alias{get_call_analysis_wdid}
\title{Return call analysis by WDID from analysis services API}
\usage{
get_call_analysis_wdid(
  wdid = NULL,
  admin_no = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  batch = FALSE,
  api_key = NULL
)
}
\arguments{
\item{wdid}{character indicating DWR unique structure identifier code (WDID)}

\item{admin_no}{character Water Right Administration Number}

\item{start_date}{character date to request data start point YYYY-MM-DD. Default is start date is "1900-01-01".}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default end date is the current date the function is run.}

\item{batch}{logical, whether to break date range calls into batches of 1 year. This can speed up data retrieval for date ranges greater than a year. A date range of 5 years would be batched into 5 separate API calls for each year. Default is FALSE, will run a single query for the entire date range.}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of call services by WDID
}
\description{
Makes a request to the analysisservices/callanalysisbywdid/ endpoint that performs a call analysis that returns a time series showing the percentage of each day that the specified WDID and priority was out of priority and the downstream call in priority.
}
