% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_frostdates.R
\name{get_climate_frostdates}
\alias{get_climate_frostdates}
\title{Return climate stations frost dates}
\usage{
get_climate_frostdates(
  station_number = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{station_number}{character, climate data station number}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of climate data frost dates data
}
\description{
Make a request to the /climatedata/climatestationfrostdates endpoint to retrieve climate stations frost dates data by station number within a given date range (start and end dates)
}
