% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_stations.R
\name{get_climate_stations}
\alias{get_climate_stations}
\title{Return Climate Station information}
\usage{
get_climate_stations(
  aoi = NULL,
  radius = NULL,
  county = NULL,
  division = NULL,
  station_name = NULL,
  site_id = NULL,
  water_district = NULL,
  api_key = NULL
)
}
\arguments{
\item{aoi}{2 column matrix/dataframe of XY coordinates, or SF point or polygon object to search for climate station structures within a given radius}

\item{radius}{numeric, search radius in miles around a given point (or the centroid of a polygon) to return administrative structures. If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}

\item{county}{character, indicating the county to query}

\item{division}{numeric, indicating the water division to query}

\item{station_name}{character, surface water station name}

\item{site_id}{character vector or list of characters of climate station site IDs}

\item{water_district}{numeric, indicating the water district to query}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of climate data station info
}
\description{
Make a request to the climatedata/climatestations/ endpoint to locate climate stations by AOI, county, division, station name, Site ID or water_district.
}
\examples{
# get climate stations in Adams County, CO
climate_stations  <- get_climate_stations(
                       county = "Adams"
                       )

# plot latitude/longitude of climate stations
plot(climate_stations$latitude~climate_stations$longitude)
}
