% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_ts.R
\name{get_climate_ts}
\alias{get_climate_ts}
\title{Return climate station timeseries data}
\usage{
get_climate_ts(
  timescale = "day",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  station_number = NULL,
  site_id = NULL,
  param = NULL,
  api_key = NULL
)
}
\arguments{
\item{timescale}{character indicating data type to return, either "day" or "month". Default is "day".}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{station_number}{character, surface water station number}

\item{site_id}{character vector or list of characters of climate station site IDs}

\item{param}{character climate variable. One of: "Evap", "FrostDate",  "MaxTemp", "MeanTemp", "MinTemp", "Precip", "Snow", "SnowDepth", "SnowSWE", "Solar","VP", "Wind"}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of climate station timeseries data
}
\description{
Make a request to the /climatedata/climatestationts  endpoints (climatestationtsday, climatestationtsmonth) to retrieve climate station timeseries data by station number or Site IDs within a given date range (start and end dates)
}
\examples{
# Retrieve daily maximum temperatures
daily_maxtemp <- get_climate_ts(
  timescale   = "day",
  start_date  = "2017-01-01",
  end_date    = "2020-01-01",
  site_id     = "USC00055984",
  param       = "MaxTemp"
  )

# plot daily maximum temp at climate station
plot(daily_maxtemp$value~daily_maxtemp$datetime, type = "l")

# Retrieve monthly precipitation
monthly_precip <- get_climate_ts(
  timescale   = "month",
  start_date  = "2000-01-01",
  end_date    = "2022-01-01",
  site_id     = "USC00055984",
  param       = "Precip"
   )

 # plot daily max temp at climate station
 plot(monthly_precip$avg_value~monthly_precip$datetime, type = "l")
}
