% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_ts.R
\name{get_climate_ts_day}
\alias{get_climate_ts_day}
\title{Return daily climate data}
\usage{
get_climate_ts_day(
  station_number = NULL,
  site_id = NULL,
  param = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{station_number}{character, climate data station number}

\item{site_id}{character vector or list of characters of climate station site IDs}

\item{param}{character climate variable. One of: "Evap", "FrostDate",  "MaxTemp", "MeanTemp", "MinTemp", "Precip", "Snow", "SnowDepth", "SnowSWE", "Solar","VP", "Wind"}

\item{start_date}{character date to request data start point YYYY-MM-DD. Default is start date is "1900-01-01".}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default end date is the current date the function is run.}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of climate data station daily time series data
}
\description{
Make a request to the /climatedata/climatestationtsday endpoint to retrieve climate stations daily time series data by station number, or Site IDs within a given date range (start and end dates)
}
