% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divrec.R
\name{get_divrec}
\alias{get_divrec}
\title{Request Diversion Records data}
\usage{
get_divrec(
  type = "divrecday",
  wdid = NULL,
  wc_identifier = "diversion",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{type}{character indicating the type of data to request. Either "divrecday", "divrecmonth", or "stagevolume". Default is "divrecday".}

\item{wdid}{character indicating WDID code of structure}

\item{wc_identifier}{character indicating whether "diversion" or "release" should be returned. Defaults to "diversion", not used in stagevolume requests.}

\item{start_date}{character date to request data start point YYYY-MM-DD.}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default is set to the current date function is run.}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with diversion records data for CDSS structure of interest
}
\description{
Make a request to CDSS API structures/divrec endpoints to get diversion records data
}
\examples{
# Request endpoint: api/v2/structures/divrec/divrecday
divrecord_day <- get_divrec(
               type             = "divrecday",
               wdid             = "2302917",
               wc_identifier    = "diversion",
               start_date       = "2000-01-01",
               end_date         = "2005-01-01"
               )

# plot daily diversion record
plot(divrecord_day$value~divrecord_day$datetime, type = "s")

# Request endpoint: api/v2/structures/divrec/divrecmonth
divrecord_month <- get_divrec(
                   type             = "divrecmonth",
                   wdid             = "2302917",
                   wc_identifier    = "diversion",
                   start_date       = "2000-01-01",
                   end_date         = "2005-01-01"
                 )

# plot monthly diversion record
plot(divrecord_month$value~divrecord_month$date, type = "s")

# Request endpoint: api/v2/structures/divrec/stagevolume
stage_vol <- get_divrec(
                   type             = "stagevolume",
                   wdid             =  "0303732",
                   start_date       = "2000-01-01",
                   end_date         = "2005-01-01"
                 )

# plot stage
plot(stage_vol$stage~stage_vol$datetime, type = "s")

# plot volume
plot(stage_vol$volume~stage_vol$datetime, type = "s")
}
