% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_groundwater}
\alias{get_groundwater}
\title{Request groundwater well measurements or search well water levels/geophysical logs}
\usage{
get_groundwater(
  type = "wellmeasurements",
  wellid = NULL,
  start_date = "1950-01-01",
  end_date = Sys.Date(),
  county = NULL,
  designated_basin = NULL,
  division = NULL,
  management_district = NULL,
  water_district = NULL,
  api_key = NULL
)
}
\arguments{
\item{type}{character indicating the type of data to search for. Either "wellmeasurements", "waterlevels", or "geophysicallogs". Default is "wellmeasurements".}

\item{wellid}{character, indicating the Well ID to query}

\item{start_date}{character date to request data start point YYYY-MM-DD.}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default is set to the current date function is run.}

\item{county}{character, indicating the county to query}

\item{designated_basin}{character, indicating the  designated basin to query}

\item{division}{numeric, indicating the water division to query}

\item{management_district}{character, indicating the management district to query}

\item{water_district}{numeric, indicating the water district to query}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with groundwater well data
}
\description{
Make a request to CDSS API /groundwater endpoints to get well measurement data or to search for groundwater wells (water level wells or geo physical log records)
}
\examples{
# Request endpoint: api/v2/groundwater/waterlevels/wellmeasurements
well_measure <- get_groundwater(
type   = "wellmeasurements",
  wellid = 1274
 )

# plot depth to water
plot(well_measure$depth_to_water~well_measure$datetime, type = "l")

# Request endpoint: api/v2/groundwater/waterlevels/wells
water_levels <- get_groundwater(
  type     = "waterlevels",
  division = 2
 )

# number of unique well IDs from query
length(unique(water_levels$well_id))

# Request endpoint: api/v2/groundwater/geophysicallogs/wells
geophysicallogs <- get_groundwater(
  type     = "geophysicallogs",
  division = 2
)

# number of unique well IDs from query
length(unique(geophysicallogs$well_id))
}
