% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_gw_gplogs_geologpicks}
\alias{get_gw_gplogs_geologpicks}
\title{Returns Groundwater Geophysical Log picks by well ID}
\usage{
get_gw_gplogs_geologpicks(wellid = NULL, api_key = NULL)
}
\arguments{
\item{wellid}{character, indicating the Well ID to query}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of groundwater geophysical log picks for a given well ID
}
\description{
Given a specified well ID, a request is made to api/v2/groundwater/geophysicallogs/geoplogpicks, and a dataframe of groundwater geophysical log picks for the given well ID is returned
}
\examples{
# Request endpoint: api/v2/groundwater/geophysicallogs/geoplogpicks/
gplogpicks <- get_gw_gplogs_geologpicks(
  wellid = 2409
 )
 gplogpicks
}
