% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_gw_gplogs_wells}
\alias{get_gw_gplogs_wells}
\title{Search for groundwater geophysicallog wells}
\usage{
get_gw_gplogs_wells(
  county = NULL,
  designated_basin = NULL,
  division = NULL,
  management_district = NULL,
  water_district = NULL,
  wellid = NULL,
  api_key = NULL
)
}
\arguments{
\item{county}{character, indicating the county to query}

\item{designated_basin}{character, indicating the  designated basin to query}

\item{division}{numeric, indicating the water division to query}

\item{management_district}{character, indicating the management district to query}

\item{water_district}{numeric, indicating the water district to query}

\item{wellid}{character, indicating the Well ID to query}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of groundwater wells within the given query specifications
}
\description{
Make a request to the groundwater/geophysicallogs/wells endpoint to retrieve groundwater geophysicallog wells data.
}
\examples{
\dontrun{
# Request endpoint: api/v2/groundwater/geophysicallogs/wells/
gplog_wells <- get_gw_gplogs_wells(
  county = "ADAMS"
 )

plot(gplog_wells$latitude~gplog_wells$longitude)
 }
}
