% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_gw_wl_wellmeasures}
\alias{get_gw_wl_wellmeasures}
\title{Request groundwater water level well measurements}
\usage{
get_gw_wl_wellmeasures(
  wellid = NULL,
  start_date = "1950-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wellid}{character, indicating the Well ID to query}

\item{start_date}{character date to request data start point YYYY-MM-DD.}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default is set to the current date function is run.}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of groundwater wells within the given query specifications
}
\description{
Given query specifications, this function makes a request to this endpoint of the CDSS API: api/v2/groundwater/waterlevels/wellmeasurements and returns a dataframe containing  groundwater water level well measurements
}
\examples{
# Request endpoint: api/v2/groundwater/waterlevels/wellmeasurements
well_measure <- get_gw_wl_wellmeasures(
                  wellid = 1274
                   )

# plot depth to water
plot(well_measure$depth_to_water~well_measure$datetime, type = "l")
}
