% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Retrieve parameter definitions for CDSS API Endpoints}
\usage{
get_params(param_url = NULL)
}
\arguments{
\item{param_url}{character. URL to CDSS API REST Help page detailing the parameter fields for each endpoint. This URL can be found in the dataframe returned by the browse_api function}
}
\value{
dataframe with the parameter name for querying, a description, and the parameter type
}
\description{
Returns the names, descriptions, and types for each parameter in a given API endpoint
}
