% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parceluse_ts.R
\name{get_parceluse_ts}
\alias{get_parceluse_ts}
\title{Return parcel use time series}
\usage{
get_parceluse_ts(
  wdid = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wdid}{character or numeric, indicating WDID code of structure}

\item{start_date}{character date to request starting year to retrieve data (YYYY-MM-DD). Default start date is "1900-01-01", therefore requests data starting from 1900.}

\item{end_date}{character date to  request ending year to retrieve data (YYYY-MM-DD). Default end date is the current date the function is run.}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of parcel use time series data
}
\description{
Make a request to the /structures/parcelusets endpoint to retrieve parcel use time series data by WDID, within a given date range (start and end dates)
}
