% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_climate_params}
\alias{get_ref_climate_params}
\title{Return climate station parameter reference table}
\usage{
get_ref_climate_params(param = NULL, api_key = NULL)
}
\arguments{
\item{param}{character, (optional) indicating the parameter to query character, if no parameter is given, all parameter dataframe is returned}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of climate station parameters
}
\description{
Make a request to the referencetables/climatestationmeastype/ endpoint to retrieve a reference table of parameters available at climate stations. The reference table functions can help provide information to use while querying other endpoints.
}
