% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_county}
\alias{get_ref_county}
\title{Return county reference table}
\usage{
get_ref_county(county = NULL, api_key = NULL)
}
\arguments{
\item{county}{character, (optional) indicating the county to query, if no county is given, entire county dataframe is returned}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of Colorado counties
}
\description{
Make a request to the referencetables/county/ endpoint to retrieve a reference table of counties. The reference table functions can help provide information to use while querying other endpoints.
}
