% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_designatedbasins}
\alias{get_ref_designatedbasins}
\title{Return designated basin reference table}
\usage{
get_ref_designatedbasins(designated_basin = NULL, api_key = NULL)
}
\arguments{
\item{designated_basin}{character, (optional) indicating the  designated basin to query character, if no designated basin is given, all designated basins dataframe is returned}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of Colorado designated basins
}
\description{
Make a request to the referencetables/designatedbasin/ endpoint to retrieve a reference table of designated basins. The reference table functions can help provide information to use while querying other endpoints.
}
