% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_divrectypes}
\alias{get_ref_divrectypes}
\title{Return Diversion Record Types reference table}
\usage{
get_ref_divrectypes(divrectype = NULL, api_key = NULL)
}
\arguments{
\item{divrectype}{character, (optional) diversion record type to query, if no divrectype is given, a dataframe with all diversion record types is returned.}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of diversion record types
}
\description{
Make a request to the referencetables/divrectypes/ endpoint to retrieve a reference table of diversion record types available. The reference table functions can help provide information to use while querying other endpoints.
}
