% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_managementdistricts}
\alias{get_ref_managementdistricts}
\title{Return management districts reference table}
\usage{
get_ref_managementdistricts(management_district = NULL, api_key = NULL)
}
\arguments{
\item{management_district}{character, (optional) indicating the management district to query, if no management district is given, dataframe of all management districts is returned}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.#' @importFrom httr GET content}
}
\value{
dataframe of Colorado management districts
}
\description{
Make a request to the referencetables/managementdistrict/ endpoint to retrieve a reference table of management districts. The reference table functions can help provide information to use while querying other endpoints.
}
