% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_stationflags}
\alias{get_ref_stationflags}
\title{Return Station Flag reference table}
\usage{
get_ref_stationflags(flag = NULL, api_key = NULL)
}
\arguments{
\item{flag}{character, (optional) short code for the flag to query, if no flag is given, a dataframe with all flags is returned.}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of station flags and flag descriptions
}
\description{
Make a request to the referencetables/stationflags/ endpoint to retrieve a reference table of station flag descriptions. The reference table functions can help provide information to use while querying other endpoints.
}
