% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_waterdistricts}
\alias{get_ref_waterdistricts}
\title{Return water districts reference table}
\usage{
get_ref_waterdistricts(division = NULL, water_district = NULL, api_key = NULL)
}
\arguments{
\item{division}{character, (optional) indicating the division to query, if no division is given, dataframe of all water districts is returned}

\item{water_district}{character, (optional) indicating the water district to query, if no ater district is given, dataframe of all water districts is returned}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of Colorado water districts, indicating the division the water district is in and its name
}
\description{
Make a request to the referencetables/waterdistrict/ endpoint to retrieve a reference table of water districts The reference table functions can help provide information to use while querying other endpoints.
}
