% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_ref_waterdivisions}
\alias{get_ref_waterdivisions}
\title{Return water divisions reference table}
\usage{
get_ref_waterdivisions(division = NULL, api_key = NULL)
}
\arguments{
\item{division}{character, (optional) indicating the division to query, if no division is given, dataframe of all water divisions is returned}

\item{api_key}{character, (optional). If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of Colorado water divisions and water division names
}
\description{
Make a request to the referencetables/waterdivision/ endpoint to retrieve a reference table of water divisions. The reference table functions can help provide information to use while querying other endpoints.
}
