% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_tbl.R
\name{get_reference_tbl}
\alias{get_reference_tbl}
\title{Return reference table information, which may help with other searches}
\usage{
get_reference_tbl(table_name = NULL, api_key = NULL)
}
\arguments{
\item{table_name}{character, indicating which reference table should be returned. One of: "county", "waterdistricts", "waterdivisions", "designatedbasins", "managementdistricts", "telemetryparams", "climateparams", "divrectypes", "flags"}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of CDSS endpoint Reference Table
}
\description{
Retrieve reference tables to help identify the correct and available inputs for generating a successful query using the other CDSS API retrieval functions in the cdssr package. This function makes requests to the /referencetables/ endpoints and returns helpful reference tables. For more information go to https://dwr.state.co.us/rest/get/help#Datasets&#ReferenceTablesController&#gettingstarted&#jsonxml
}
\examples{
# Retrieve station flag reference table
flag_tbl <- get_reference_tbl(
   table_name = "flags"
   )
# Retrieve available telemetry station parameters
telemetry_params <- get_reference_tbl(
   table_name = "telemetryparams"
   )
# Retrieve available climate station parameters
climate_params <- get_reference_tbl(
   table_name = "climateparams"
   )
# Retrieve water districts
water_districts <- get_reference_tbl(
   table_name = "waterdistricts"
   )
}
