% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysisservices.R
\name{get_source_route_framework}
\alias{get_source_route_framework}
\title{Return DWR source route framework from analysis services API}
\usage{
get_source_route_framework(
  division = NULL,
  gnis_name = NULL,
  water_district = NULL,
  api_key = NULL
)
}
\arguments{
\item{division}{character, the division to query and retrieve DWR source route frameworks. Defaults to NULL.}

\item{gnis_name}{character, the GNIS Name to query and retrieve DWR source route frameworks. Defaults to NULL.}

\item{water_district}{character, the water district to query and retrieve DWR source route frameworks. Defaults to NULL.}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of water source route framework
}
\description{
Makes a request to the analysisservices/watersourcerouteframework/ endpoint to retrieve the DWR source route framework reference table for the criteria specified.
}
