% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_divrecday.R
\name{get_structure_divrecday}
\alias{get_structure_divrecday}
\title{Request Structure Daily Diversion/Release Records}
\usage{
get_structure_divrecday(
  wdid = NULL,
  wc_identifier = "diversion",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wdid}{character or numeric indicating WDID code of structure}

\item{wc_identifier}{character indicating whether "diversion" or "release" should be returned. Defaults to "diversion"}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with diversion or flow data for CDSS structure of interest
}
\description{
Request Structure Daily Diversion/Release Records
}
\examples{
# Retrieve daily diversion records for given period
daily_div <- get_structure_divrecday(
               wdid             = "2302917",
               wc_identifier    = "diversion",
               start_date       = "1988-01-01",
               end_date         = "1992-01-01"
           )
daily_div

# plot daily diversions
plot(daily_div$value~daily_div$datetime, type = "s")
}
