% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_divrecmonth.R
\name{get_structure_divrecmonth}
\alias{get_structure_divrecmonth}
\title{Request Structure Monthly Diversion/Release Records}
\usage{
get_structure_divrecmonth(
  wdid = NULL,
  wc_identifier = "diversion",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wdid}{character or numeric indicating WDID code of structure}

\item{wc_identifier}{character indicating whether "diversion" or "release" should be returned. Defaults to "diversion"}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with monthly structure data for the CDSS structure of interest
}
\description{
Request Structure Monthly Diversion/Release Records
}
