% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure_stage.R
\name{get_structure_stage}
\alias{get_structure_stage}
\title{Request Structure stage/volume Records}
\usage{
get_structure_stage(
  wdid = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wdid}{character indicating WDID code of structure}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with stage/volume data for CDSS structure of interest
}
\description{
Request Structure stage/volume Records
}
