% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structures_divrec.R
\name{get_structures_divrec}
\alias{get_structures_divrec}
\title{Request Diversion Records data for administrative structures}
\usage{
get_structures_divrec(
  wdid = NULL,
  wc_identifier = "diversion",
  type = "day",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{wdid}{character vector or list of characters indicating WDID code of structure. Only a single WDID can be used when type = "stagevolume".}

\item{wc_identifier}{character indicating whether "diversion" or "release" should be returned. A wc_identifier does not need to be given in stagevolume requests. Defaults to "diversion".}

\item{type}{character indicating the type of data to request. Either "day", "month", "year", or "stagevolume". Default is "day".}

\item{start_date}{character date to request data start point YYYY-MM-DD. Default is start date is "1900-01-01".}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default end date is the current date the function is run.}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with diversion records data for CDSS structure of interest
}
\description{
Make a request to CDSS API structures/divrec endpoints to get diversion/releases/stage/volume data for administrative structures. For more information go here https://dwr.state.co.us/rest/get/help#Datasets&#DiversionRecordsController&https://dnrweblink.state.co.us/dwr/ElectronicFile.aspx?docid=3600965&dbid=0&#gettingstarted&#jsonxml
}
\examples{
# Request endpoint: api/v2/structures/divrec/divrecday
divrecord_day <- get_structures_divrec(
               wdid             = "2302917",
               type             = "day",
               wc_identifier    = "diversion",
               start_date       = "2000-01-01",
               end_date         = "2005-01-01"
               )

# plot daily diversion record
plot(divrecord_day$data_value~divrecord_day$datetime, type = "s")

# Request endpoint: api/v2/structures/divrec/divrecmonth
divrecord_month <- get_structures_divrec(
                   wdid             = "2302917",
                   type             = "month",
                   wc_identifier    = "diversion",
                   start_date       = "2000-01-01",
                   end_date         = "2005-01-01"
                 )

# plot monthly diversion record
plot(divrecord_month$data_value~divrecord_month$datetime, type = "s")

# Request endpoint: api/v2/structures/divrec/divrecyear
divrecord_year <- get_structures_divrec(
 wdid             = "2302917",
 type             = "year",
 wc_identifier    = "diversion",
 start_date       = "2000-01-01",
 end_date         = "2022-01-01"
 )

# plot yearly diversion record
plot(divrecord_year$data_value~divrecord_year$data_meas_date, type = "s")

# Request endpoint: api/v2/structures/divrec/stagevolume
stage_vol <- get_structures_divrec(
                   wdid             =  "0303732",
                   type             = "stagevolume",
                   start_date       = "2000-01-01",
                   end_date         = "2005-01-01"
                 )

# plot stage
plot(stage_vol$stage~stage_vol$datetime, type = "s")

# plot volume
plot(stage_vol$volume~stage_vol$datetime, type = "s")
}
