% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structures_divrec_ts.R
\name{get_structures_divrec_ts}
\alias{get_structures_divrec_ts}
\title{Return diversion/releases record data for administrative structures}
\usage{
get_structures_divrec_ts(
  wdid = NULL,
  wc_identifier = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  timescale = NULL,
  api_key = NULL
)
}
\arguments{
\item{wdid}{character vector or list of characters indicating WDID code of structure.}

\item{wc_identifier}{character indicating whether "diversion" or "release" should be returned. Default is NULL which will return diversions.}

\item{start_date}{character date to request data start point YYYY-MM-DD. Default is start date is "1900-01-01".}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default end date is the current date the function is run.}

\item{timescale}{character indicating the time series time step. Either "day", "month", "year". Default is to return daily time series.}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with diversion records data for CDSS structure of interest
}
\description{
Make a request to the CDSS API /structures/divrec endpoints to get diversion/releases time series data for administrative structures by wdid, within a given date range (start and end dates) and at a specified temporal resolution.
}
\examples{
\dontrun{
# Request endpoint: api/v2/structures/divrec/divrecday
divrecord_day <- get_structures_divrec_ts(
               wdid             = "2302917",
               wc_identifier    = "diversion",
               start_date       = "2000-01-01",
               end_date         = "2005-01-01",
               timescale        = "day"
               )

# plot daily diversion record
plot(divrecord_day$data_value~divrecord_day$datetime, type = "s")

# Request endpoint: api/v2/structures/divrec/divrecmonth
divrecord_month <- get_structures_divrec_ts(
                   wdid             = "2302917",
                   wc_identifier    = "diversion",
                   start_date       = "2000-01-01",
                   end_date         = "2005-01-01",
                   timescale        = "month"
                 )

# plot monthly diversion record
plot(divrecord_month$data_value~divrecord_month$datetime, type = "s")

# Request endpoint: api/v2/structures/divrec/divrecyear
divrecord_year <- get_structures_divrec_ts(
                    wdid             = "2302917",
                    wc_identifier    = "diversion",
                    start_date       = "2000-01-01",
                    end_date         = "2022-01-01",
                    timescale        = "year"
                    )

# plot yearly diversion record
plot(divrecord_year$data_value~divrecord_year$data_meas_date, type = "s")
}
}
