% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sw_stations.R
\name{get_sw_stations}
\alias{get_sw_stations}
\title{Return Surface Water Station information}
\usage{
get_sw_stations(
  aoi = NULL,
  radius = NULL,
  abbrev = NULL,
  county = NULL,
  division = NULL,
  station_name = NULL,
  usgs_id = NULL,
  water_district = NULL,
  api_key = NULL
)
}
\arguments{
\item{aoi}{list of length 2 containing an XY coordinate pair, 2 column matrix/dataframe of XY coordinates, sf or Terra SpatVector point/polygon/linestring geometry}

\item{radius}{numeric, search radius in miles around given point (or the centroid of a polygon). If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}

\item{abbrev}{character vector or list of characters of station abbreviation}

\item{county}{character, indicating the county to query}

\item{division}{numeric, indicating the water division to query}

\item{station_name}{character, surface water station name}

\item{usgs_id}{character vector or list of characters of USGS Site IDs}

\item{water_district}{numeric, indicating the water district to query}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of surface water station info
}
\description{
Make a request to the /surfacewater/surfacewaterstations endpoint to locate surface water stations by AOI, station abbreviation, county, division, station name, USGS ID or water_district.
}
\examples{
\dontrun{
# Retrieve surface water station info from Larimer county
sw_stations <- get_sw_stations(
                    county = "Larimer"
                    )
# plot latitude/longitude of surface water stations
plot(sw_stations$latitude~sw_stations$longitude)
}
}
