% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sw_ts.R
\name{get_sw_ts_wyear}
\alias{get_sw_ts_wyear}
\title{Return water year surface water time series data}
\usage{
get_sw_ts_wyear(
  abbrev = NULL,
  station_number = NULL,
  usgs_id = NULL,
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  api_key = NULL
)
}
\arguments{
\item{abbrev}{character vector or list of characters of station abbreviation}

\item{station_number}{character, surface water station number}

\item{usgs_id}{character vector or list of characters of USGS Site IDs}

\item{start_date}{character date to request data start point YYYY-MM-DD. Default is start date is "1900-01-01".}

\item{end_date}{character date to request data end point YYYY-MM-DD. Default end date is the current date the function is run.}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of annual surface water station time series data
}
\description{
Make a request to the /surfacewater/surfacewatertswateryear endpoint to retrieve surface water stations water year time series data by station abbreviations, station number, or USGS Site IDs within a given date range (start and end dates)
}
