% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telemetry_stations.R
\name{get_telemetry_stations}
\alias{get_telemetry_stations}
\title{Return Telemetry Station information}
\usage{
get_telemetry_stations(
  aoi = NULL,
  radius = NULL,
  abbrev = NULL,
  county = NULL,
  division = NULL,
  gnis_id = NULL,
  usgs_id = NULL,
  water_district = NULL,
  wdid = NULL,
  api_key = NULL
)
}
\arguments{
\item{aoi}{list of length 2 containing an XY coordinate pair, 2 column matrix/dataframe of XY coordinates, sf or Terra SpatVector point/polygon/linestring geometry}

\item{radius}{numeric, search radius in miles around given point (or the centroid of a polygon). If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}

\item{abbrev}{character vector or list of characters of station abbreviation}

\item{county}{character, indicating the county to query}

\item{division}{numeric, indicating the water division to query}

\item{gnis_id}{character, water source - Geographic Name Information System ID}

\item{usgs_id}{character, indicating USGS Station ID}

\item{water_district}{numeric, indicating the water district to query}

\item{wdid}{character indicating WDID code of structure}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of telemetry station info
}
\description{
Make a request to the telemetrystations/telemetrystation/ endpoint to locate telemetry stations by AOI, station abbreviations, county, division, station abbreviation, GNIS ID, USGS Station ID, or WDID
}
\examples{
\dontrun{
# Retrieve telemetry stations within a county
telemetry_station <- get_telemetry_stations(
                               county = "Adams"
                               )

# plot telemetry station locations
plot(telemetry_station$latitude~telemetry_station$longitude)
}
}
