% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_telemetry_ts.R
\name{get_telemetry_ts}
\alias{get_telemetry_ts}
\title{Request Telemetry station timeseries data}
\usage{
get_telemetry_ts(
  abbrev = NULL,
  parameter = "DISCHRG",
  start_date = "1900-01-01",
  end_date = Sys.Date(),
  include_third_party = TRUE,
  timescale = "day",
  api_key = NULL
)
}
\arguments{
\item{abbrev}{character indicating station abbreviation}

\item{parameter}{character indicating which parameter should be retrieved. Default is "DISCHRG" (discharge), all parameters are not avaliable at all telemetry stations.}

\item{start_date}{character date to request data start point YYYY-MM-DD}

\item{end_date}{character date to request data end point YYYY-MM-DD}

\item{include_third_party}{logical, Whether to retrieve data from other third party sources if necessary. Default is TRUE}

\item{timescale}{character indicating data type to return, either "raw", "hour", or "day". Default is "day"}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe with telemetry station timeseries data
}
\description{
Request Telemetry station timeseries data
}
\examples{
# Retrieve daily discharge for CLAFTCCO telemetry station
ts <- get_telemetry_ts(
           abbrev              = "CLAFTCCO",
           parameter           = "DISCHRG",
           start_date          = "2015-01-01",
           end_date            = "2022-01-01",
           timescale           = "day",
           include_third_party = TRUE
           )
ts

# Plot daily discharge data
plot(ts$value~ts$datetime, type = "l")
}
