% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_rights.R
\name{get_water_rights_netamount}
\alias{get_water_rights_netamount}
\title{Return water rights net amounts data
Returns current status of a water right based on all of its court decreed actions from the CDSS API waterrights/netamount endpoint.}
\usage{
get_water_rights_netamount(
  county = NULL,
  division = NULL,
  water_district = NULL,
  wdid = NULL,
  api_key = NULL
)
}
\arguments{
\item{county}{character, indicating the county to query}

\item{division}{numeric, indicating the water division to query}

\item{water_district}{numeric, indicating the water district to query}

\item{wdid}{character, indicating WDID code of structure}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of water right net amounts
}
\description{
Return water rights net amounts data
Returns current status of a water right based on all of its court decreed actions from the CDSS API waterrights/netamount endpoint.
}
\examples{
# get_water_rights_netamount(county = "Adams")
# get_water_rights_netamount(division = 1)
# get_water_rights_netamount(wdid   = "0100555")
}
