% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_rights.R
\name{get_water_rights_trans}
\alias{get_water_rights_trans}
\title{Return water rights transactions data}
\usage{
get_water_rights_trans(
  aoi = NULL,
  radius = NULL,
  county = NULL,
  division = NULL,
  water_district = NULL,
  wdid = NULL,
  api_key = NULL
)
}
\arguments{
\item{aoi}{list of length 2 containing an XY coordinate pair, 2 column matrix/dataframe of XY coordinates, sf or Terra SpatVector point/polygon/linestring geometry}

\item{radius}{numeric, search radius in miles around given point (or the centroid of a polygon). If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}

\item{county}{character, indicating the county to query}

\item{division}{numeric, indicating the water division to query}

\item{water_district}{numeric, indicating the water district to query}

\item{wdid}{character, indicating WDID code of structure}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of water right transactions
}
\description{
Make a request to the /waterrights/transaction endpoint to retrieve water rights transactions data via a spatial search or by county, division, water district, or WDID, within a given date range (start and end dates). Returns List of court decreed actions that affect amount and use(s) that can be used by each water right.
}
\examples{
\dontrun{
# Water right transactions within a county
county_wr <- get_water_rights_trans(county = "Adams")


# Water right transactions within a division
division_wr <- get_water_rights_trans(division = 1)


# Water right transactions for a WDID
wdid_wr <- get_water_rights_trans(wdid   = "2000502")


# Water right transactions within a 20 mile search radius
aoi_wr <- get_water_rights_trans(
                  aoi    = data.frame(X = -104.3379, Y = 39.87417),
                  radius = 20
                  )
}
}
