% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_convert}
\alias{null_convert}
\title{Convert non NULL list/vectors to characters}
\usage{
null_convert(x)
}
\arguments{
\item{x}{list, or vector to convert to character}
}
\value{
character list/vector of the same length as x
}
\description{
Internal function for converting non NULL lists/vectors to characters. When NULL is provided, NULL is returned.
}
