% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_date}
\alias{parse_date}
\title{Formats, collapses, and extract dates to build query strings}
\usage{
parse_date(date = NULL, start = TRUE, format = "\%m-\%d-\%Y", sep = "\%2F")
}
\arguments{
\item{date}{character date in YYYY-MM-DD format.}

\item{start}{logical, whether the given date is the starting or ending date}

\item{format}{character, format to convert the date to. Defaults to MM-DD-YYYY.}

\item{sep}{character separator to collapse date by. Defaults to "\%2F"}
}
\value{
reformated date string
}
\description{
Internal function for generating date query string URls
}
