% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_gets}
\alias{parse_gets}
\title{Parse GET URL response}
\usage{
parse_gets(url = NULL)
}
\arguments{
\item{url}{character URL if page to retrieve. Default is NULL.}
}
\value{
json formatted list
}
\description{
Internal convenience function for parsing a "text" content of URL GET responses into JSON format
}
