% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_endpoint.R
\name{preview_endpoint}
\alias{preview_endpoint}
\title{Retrieve descriptions of expected return fields for an API endpoint}
\usage{
preview_endpoint(endpoint_url = NULL)
}
\arguments{
\item{endpoint_url}{character. URL to CDSS API REST Help page detailing the return fields for each endpoint. This URL can be found in the dataframe returned by the browse_api function}
}
\value{
dataframe with the field name, a description, and the data type
}
\description{
Returns the names, descriptions, and types for each response field for a given API endpoint
}
