% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_error}
\alias{query_error}
\title{GET Request Error message handler}
\usage{
query_error(arg_lst = NULL, url = NULL, ignore = NULL, e_msg = NULL)
}
\arguments{
\item{arg_lst}{list of function arguments by calling 'as.list(environment())'. Default is NULL.}

\item{url}{character URL if page to retrieve. Default is NULL.}

\item{ignore}{character vector of function arguments to ignore NULL check. Default is NULL.}

\item{e_msg}{character error message. Default is NULL.}
}
\value{
character error message that includes the query inputs that led to the error, the requested URL, and the original error message
}
\description{
Internal function for generating dynamic error messages for failed GET requests. Designed to be called within another function and print out the functions input arguments.
}
