% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_meta_data.R
\docType{data}
\name{resource_meta}
\alias{resource_meta}
\title{Meta data for CDSS REST API resources}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1031 rows and 5 columns.
}
\usage{
resource_meta
}
\description{
Dataframe containing meta data for the data returned by CDSS REST API endpoints
}
\examples{
\dontrun{
 cdssr::resource_meta
}

}
\keyword{datasets}
