% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_args}
\alias{str_args}
\title{Convert function arguments to strings}
\usage{
str_args(arg_lst = NULL, ignore = NULL, envir)
}
\arguments{
\item{arg_lst}{list of function arguments by calling 'as.list(environment())'}

\item{ignore}{character vector of function arguments to ignore NULL check}

\item{envir}{an environment. If called within another function to alter the arguments of the outer function, use envir = environment()}
}
\value{
No returns, function reassigns function arguments to strings
}
\description{
Internal function for converting all or specific function arguments to strings
}
