% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_radius}
\alias{check_radius}
\title{Validate radius for location search queries}
\usage{
check_radius(aoi, radius, verbose = FALSE)
}
\arguments{
\item{aoi}{2 column matrix/dataframe of XY coordinates, or SF point or polygon object to search for administrative structures within a given radius}

\item{radius}{numeric, search radius in miles around a given point (or the centroid of a polygon) to return administrative structures. If an AOI is given, radius defaults to 20 miles. If no AOI is given, then default is NULL.}

\item{verbose}{logical, whether to print messages to console. Default is FALSE, no print.}
}
\value{
numeric radius value or NULL
}
\description{
Internal helper function. Checks that an AOI is given and that the radius is within valid range of values. Radius must be: 0 < radius < 150.
}
