% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_groundwater_well_search}
\alias{get_groundwater_well_search}
\title{Search groundwater wells (water levels or geo physical logs)}
\usage{
get_groundwater_well_search(
  search = "waterlevels",
  division = NULL,
  county = NULL,
  designated_basin = NULL,
  water_district = NULL,
  management_district = NULL,
  wellid = NULL,
  api_key = NULL
)
}
\arguments{
\item{search}{character indicating which endpoint to use for search. Either "waterlevels" or "geophysicallogs". Default is "waterlevels".}

\item{division}{numeric, indicating the water division to query}

\item{county}{character, indicating the county to query}

\item{designated_basin}{character, indicating the  designated basin to query}

\item{water_district}{numeric, indicating the water district to query}

\item{management_district}{character, indicating the management district to query}

\item{wellid}{numeric, indicating the Well ID to query}

\item{api_key}{character, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of groundwater wells within the given query specifications
}
\description{
Given the "search" parameter, a request is made to these endpoints api/v2/groundwater/waterlevels/wells or  api/v2/groundwater/geophysicallogs/wells, and a dataframe of groundwater wells is returned matching the specified query
}
