% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groundwater.R
\name{get_gw_gplogs_geologpicks}
\alias{get_gw_gplogs_geologpicks}
\title{Return Groundwater Geophysical Log picks by well ID}
\usage{
get_gw_gplogs_geologpicks(wellid = NULL, api_key = NULL)
}
\arguments{
\item{wellid}{character, indicating the Well ID to query}

\item{api_key}{character, API authorization token, optional. If more than maximum number of requests per day is desired, an API key can be obtained from CDSS.}
}
\value{
dataframe of groundwater geophysical log picks for a given well ID
}
\description{
Make a request to the groundwater/geophysicallogs/wells endpoint to retrieve groundwater geophysical log picks for the given well ID.
}
\examples{
\dontrun{
# Request endpoint: api/v2/groundwater/geophysicallogs/geoplogpicks/
gplogpicks <- get_gw_gplogs_geologpicks(
  wellid = 2409
 )
}
}
